feat: Address warnings from clang-tidy  (#78)
This PR introduces `clang-tidy` and applies a large number of fixes it
suggested.

The changes include:
- style modernizations, e.g., using `_v` and `_t` suffixes, `using`
instead of `typedef`, range-based for loops
- perf improvements, e.g., using `std::move`
- and correctness fixes, e.g., adding `noexcept`, explicit casts, fixing
dangling pointers

Note that `clang-tidy` is not added to CI in this PR, and to manually
trigger trigger clang-tidy in this codebase, we will have to run:

```
uv run --with "clang-tidy==21.1.1"          \
  python tests/lint/clang_tidy_precommit.py \
    --build-dir=build-pre-commit            \
    --jobs=16                               \
    ./src/ ./include ./tests
```

+++ src/ffi/container.cc
@@ -79,7 +79,9 @@ TVM_FFI_STATIC_INIT_BLOCK() {
            [](const ffi::MapObj* n) -> int64_t { return static_cast<int64_t>(n->size()); })
       .def("ffi.MapGetItem", [](const ffi::MapObj* n, const Any& k) -> Any { return n->at(k); })
       .def("ffi.MapCount",
-           [](const ffi::MapObj* n, const Any& k) -> int64_t { return n->count(k); })
+           [](const ffi::MapObj* n, const Any& k) -> int64_t {
+             return static_cast<int64_t>(n->count(k));
+           })
       .def("ffi.MapForwardIterFunctor", [](const ffi::MapObj* n) -> ffi::Function {
         return ffi::Function::FromTyped(MapForwardIterFunctor(n->begin(), n->end()));
       });
